+++ Core/Casa.Application/Models/ManageTicketModel.cs
@@ -446,9 +446,9 @@ namespace CLN.Application.Models
 
         public int? SpareDetailsId { get; set; }
 
-        public int? Quantity { get; set; }
+        public decimal? Quantity { get; set; }
 
-        public int? AvailableQty { get; set; }
+        public decimal? AvailableQty { get; set; }
 
         //public int? PartStatusId { get; set; }
     }
@@ -1479,9 +1479,9 @@ namespace CLN.Application.Models
 
         public string? SpareDesc { get; set; }
 
-        public int? Quantity { get; set; }
+        public decimal? Quantity { get; set; }
 
-        public int? AvailableQty { get; set; }
+        public decimal? AvailableQty { get; set; }
 
         //public int? PartStatusId { get; set; }
